Rejig methods and only load assets on non-api requests
